Renamed blocks to block and types to minecraft_types
